# Speedup MIN and MAX for native types