… src/Common/TargetSpecific.h
… tests/queries/0_stateless/02406_minmax_behaviour.reference
… tests/queries/0_stateless/02406_minmax_behaviour.sql