### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
- Speedup MIN and MAX for native types


Continuation from https://github.com/ClickHouse/ClickHouse/pull/40633 but instead of implementing new classes, which was deemed too dangerous because it duplicated storage code, it implements the new performance improvements by declaring subclasses that implement `addBatchSinglePlace` and `addBatchSinglePlaceNotNull` for min and max.


Note that I've only enabled SSE and AVX2 because I didn't see any improvement for AVX512 instructions, so less code to compile.

What's missing from the original PR but I intent to push in different PRs:
* Do a similar improvement change `ANY` to remove the `AggregateFunctionsSingleValue::is_any` hack.
* Similar improvements for ARGMIN / ARGMAX and its combinator (by bringing `findNumericExtremeIndex`)
* Improvements in MIN/MAX for generic types (e.g. tuple) by using column.compareAt.