@Algunenano 2023-12-27
I've been doing some local tests and it does show some improvement, but way less than expected. The problem with my assumption is that the if aggregate function execution was 20% of the query you can't speed it up over 20% :smile: 


An extract from perf:
```
+    1.40%     1.29%  QueryPipelineEx  clickhouse                                      [.] std::__1::optional<unsigned short> DB::findNumericExtremeImplAVX2<unsigned short, DB::MinComparator<unsigned short>, true, false>(unsigned short const*, char8_t ▒
+    1.22%     0.64%  QueryPipelineEx  clickhouse                                      [.] MemoryTracker::getSampleProbability(unsigned long)                                                                                                               ▒
     1.19%     0.86%  QueryPipelineEx  clickhouse                                      [.] DB::injection(double, double, double, double)                                                                                                                    ▒
+    1.18%     0.06%  QueryPipelineEx  [kernel.vmlinux]                                [k] wake_up_q                                                                                                                                                        ▒
+    1.11%     0.51%  QueryPipelineEx  [kernel.vmlinux]                                [k] try_to_wake_up                                                                                                                                                   ▒
+    1.11%     0.35%  QueryPipelineEx  clickhouse                                      [.] ProfileEvents::increment(StrongTypedef<unsigned long, ProfileEvents::EventTag>, unsigned long)                                                                   ▒
     1.09%     0.71%  QueryPipelineEx  clickhouse                                      [.] operator delete(void*, unsigned long)                                                                                                                            ▒
+    1.03%     0.35%  QueryPipelineEx  clickhouse                                      [.] DB::ISimpleTransform::prepare()
```

Now it's spending pretty much the same amount of time calculating the min as in `MemoryTracker::getSampleProbability` (which is off) or `ProfileEvents::increment`.