@Algunenano 2023-12-27
Some perf tests:

Setup:
```
create table memory_numbers (u8 UInt8, i64 Int64, f32 Float32, f64 Float64, nu8 Nullable(UInt8), ni64 Nullable(Int64), nf32 Nullable(Float32), nf64 Nullable(Float64)) ENGINE=Memory;
INSERT INTO memory_numbers SELECT
    number,
    number,
    number,
    number,
    if(((number % 3) = 0) OR ((number % 5) = 1), NULL, number),
    if(((number % 3) = 0) OR ((number % 5) = 1), NULL, number),
    if(((number % 3) = 0) OR ((number % 5) = 1), NULL, number),
    if(((number % 3) = 0) OR ((number % 5) = 1), NULL, number)
FROM numbers_mt(1000000000);
```

### MIN

```
for i in $(printf "u8\ni64\nf32\nf64\nnu8\nni64\nnf32\nnf64"); do echo $i; clickhouse benchmark --port 49000 --timelimit 10 -q "Select min($i) from memory_numbers; "; echo "^^ $i ^^"; done
```


* u8
  * Before: localhost:49000, queries: 363, QPS: 36.114, RPS: 36114356350.544, MiB/s: 34441.334, result RPS: 36.114, result MiB/s: 0.000.
  * After: localhost:49000, queries: 395, QPS: 39.409, RPS: 39408625340.607, MiB/s: 37582.994, result RPS: 39.409, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 405, QPS: 40.311, RPS: 40311354466.735, MiB/s: 38443.903, result RPS: 40.311, result MiB/s: 0.000.
* i64
  * Before: localhost:49000, queries: 66, QPS: 6.532, RPS: 6532466086.540, MiB/s: 49838.761, result RPS: 6.532, result MiB/s: 0.000.
  * After: localhost:49000, queries: 67, QPS: 6.648, RPS: 6647547599.454, MiB/s: 50716.763, result RPS: 6.648, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 66, QPS: 6.555, RPS: 6554882065.039, MiB/s: 50009.781, result RPS: 6.555, result MiB/s: 0.000.
* f32
  * Before: localhost:49000, queries: 130, QPS: 12.831, RPS: 12831239058.107, MiB/s: 48947.293, result RPS: 12.831, result MiB/s: 0.000.
  * After: localhost:49000, queries: 132, QPS: 13.084, RPS: 13083748605.294, MiB/s: 49910.540, result RPS: 13.084, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 130, QPS: 12.846, RPS: 12846424404.402, MiB/s: 49005.220, result RPS: 12.846, result MiB/s: 0.000.
* f64
  * Before: localhost:49000, queries: 66, QPS: 6.520, RPS: 6519767205.583, MiB/s: 49741.876, result RPS: 6.520, result MiB/s: 0.000.
  * After: localhost:49000, queries: 68, QPS: 6.711, RPS: 6710618523.220, MiB/s: 51197.956, result RPS: 6.711, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 66, QPS: 6.546, RPS: 6546223501.223, MiB/s: 49943.722, result RPS: 6.546, result MiB/s: 0.000.
* nullable u8
  * Before: localhost:49000, queries: 239, QPS: 23.720, RPS: 23719665481.669, MiB/s: 45241.672, result RPS: 23.720, result MiB/s: 0.000.
  * After: localhost:49000, queries: 240, QPS: 23.886, RPS: 23886213623.950, MiB/s: 45559.337, result RPS: 23.886, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 237, QPS: 23.563, RPS: 23562716521.170, MiB/s: 44942.315, result RPS: 23.563, result MiB/s: 0.000.
* nullable i64
  * Before: localhost:49000, queries: 60, QPS: 5.863, RPS: 5863340817.145, MiB/s: 50325.458, result RPS: 5.863, result MiB/s: 0.000.
  * After: localhost:49000, queries: 62, QPS: 6.095, RPS: 6094690408.795, MiB/s: 52311.147, result RPS: 6.095, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 59, QPS: 5.798, RPS: 5797674669.685, MiB/s: 49761.841, result RPS: 5.798, result MiB/s: 0.000.
* nullable f32
  * Before: localhost:49000, queries: 103, QPS: 10.115, RPS: 10114767729.566, MiB/s: 48230.971, result RPS: 10.115, result MiB/s: 0.000.
  * After: localhost:49000, queries: 107, QPS: 10.592, RPS: 10592291145.637, MiB/s: 50507.980, result RPS: 10.592, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 104, QPS: 10.237, RPS: 10237331973.631, MiB/s: 48815.403, result RPS: 10.237, result MiB/s: 0.000.
* nullable f64
  * Before: localhost:49000, queries: 58, QPS: 5.734, RPS: 5733669427.709, MiB/s: 49212.479, result RPS: 5.734, result MiB/s: 0.000.
  * After: localhost:49000, queries: 62, QPS: 6.088, RPS: 6088118360.913, MiB/s: 52254.739, result RPS: 6.088, result MiB/s: 0.000.
  * JIT: localhost:49000, queries: 59, QPS: 5.830, RPS: 5829798633.458, MiB/s: 50037.563, result RPS: 5.830, result MiB/s: 0.000.


Some thoughts:
* The impact of the agg function in these queries is minimal, so a great improvement in exectution of the function only improves the query marginally.
* I run this in a CPU with heterogeneous cache sizes and turbo boost activated. Not a great idea.
* The new functions are always faster than the old. Almost always also faster than JIT. Should we just remove it for those functions?